refactor(jj-test-fixtures): split into cmd + per-fixture modules
- src/cmd.rs: jj command runners (run_jj_in, init_colocated,
  configure_test_user) — reusable across all fixtures
- src/linear.rs: LinearFixture — uses the cmd helpers, owns its own
  file writes
- src/lib.rs: thin module index + flat re-exports so callers can
  still write `jj_test_fixtures::LinearFixture` / `run_jj_in`

Future fixtures (conflict, divergent) drop in as siblings of linear.rs
without touching cmd helpers.
